Fix update rate issues by working around MutliThreadedExecutor
Currently the MutliThreadedExecutor performance is very bad. This leads
to controllers not meeting their update rate. This PR is a temporary
workaround for these issues.

The current approach uses a `rclcpp` timer to execute the control loop.
When used in combination with the `MutliThreadedExecutor`, the timers
are not execute at their target frequency. I've converted the control
loop to a while loop on a separate thread that uses `nanosleep` to
execute the correct update rate. This means that `rclcpp` is not
involved in the execution and leads to much better performance.