Resolve deadlock in calling controller_manager services
Apparently, sercice calls can end up in a deadlock where the server says
it cannot send the response to the client. In that case, if we spin_until_future_complete()
we will be in a deadlock forever. Hence, this commit adds a timeout and retry
mechanism to the service call abstraction.